+++ opentelemetry-sdk/src/export/logs/mod.rs
@@ -12,9 +12,9 @@ use std::{borrow::Cow, fmt::Debug};
 /// `LogExporter` defines the interface that log exporters should implement.
 #[async_trait]
 pub trait LogExporter: Send + Debug {
-    /// Exports a batch of `ResourceLogs`.
+    /// Exports a batch of [`LogData`].
     async fn export(&mut self, batch: Vec<LogData>) -> LogResult<()>;
-    /// Shuts down the expoter.
+    /// Shuts down the exporter.
     fn shutdown(&mut self) {}
     #[cfg(feature = "logs_level_enabled")]
     /// Chek if logs are enabled.